requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.2.1
django-ckeditor==6.7.2
django-hide-admin==1.0.0
django-js-asset==3.1.2
django-lockdown==5.0.0
django-redis==5.4.0
django-unfold==0.60.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.3
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycountry==24.6.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
